Move model tests into separate files
